+++ lib/Api/PaymentContractsApi.php
@@ -893,4 +893,100 @@ class PaymentContractsApi
             }
         }
     }
+
+    /**
+     * Operation revokeAccrual
+     *
+     * POST Payment/Contracts/{paymentContractId}/revokeAccrual
+     *
+     * @param string $payment_contract_id Contract identifier (required)
+     * @throws ApiException on non-2xx response
+     * @return void
+     */
+    public function revokeAccrual($payment_contract_id)
+    {
+        $this->revokeAccrualWithHttpInfo($payment_contract_id);
+    }
+
+    /**
+     * Operation revokeAccrualWithHttpInfo
+     *
+     * POST Payment/Contracts/{paymentContractId}/revokeAccrual
+     *
+     * @param string $payment_contract_id Contract identifier (required)
+     * @throws ApiException on non-2xx response
+     * @return array of null, HTTP status code, HTTP response headers (array of strings)
+     */
+    public function revokeAccrualWithHttpInfo($payment_contract_id)
+    {
+        // verify the required parameter 'payment_contract_id' is set
+        if ($payment_contract_id === null || (is_array($payment_contract_id) && count($payment_contract_id) === 0)) {
+            throw new \InvalidArgumentException(
+                'Missing the required parameter $payment_contract_id when calling revokeAccrual'
+            );
+        }
+        // parse inputs
+        $resourcePath = "/Payment/Contracts/{paymentContractId}/revokeAccrual";
+        $httpBody = '';
+        $queryParams = [];
+        $headerParams = [];
+        $formParams = [];
+        $_header_accept = $this->apiClient->selectHeaderAccept(['application/json']);
+        if (!is_null($_header_accept)) {
+            $headerParams['Accept'] = $_header_accept;
+        }
+        $headerParams['Content-Type'] = $this->apiClient->selectHeaderContentType([]);
+
+        // path params
+        if ($payment_contract_id !== null) {
+            $resourcePath = str_replace(
+                "{" . "paymentContractId" . "}",
+                $this->apiClient->getSerializer()->toPathValue($payment_contract_id),
+                $resourcePath
+            );
+        }
+
+        // for model (json/xml)
+        if (isset($_tempBody)) {
+            $httpBody = $_tempBody; // $_tempBody is the method argument, if present
+        } elseif (count($formParams) > 0) {
+            $httpBody = $formParams; // for HTTP post (form)
+        }
+        for ($retries = 0; ; $retries++) {
+
+            // this endpoint requires OAuth (access token)
+            if (strlen($this->apiClient->getConfig()->getAccessToken()) !== 0) {
+                $headerParams['Authorization'] = 'Bearer ' . $this->apiClient->getConfig()->getAccessToken();
+            }
+
+            // make the API Call
+            try {
+                list($response, $statusCode, $httpHeader) = $this->apiClient->callApi(
+                    $resourcePath,
+                    'POST',
+                    $queryParams,
+                    $httpBody,
+                    $headerParams,
+                    null,
+                    '/Payment/Contracts/{paymentContractId}/revokeAccrual'
+                );
+
+                return [null, $statusCode, $httpHeader];
+            } catch (ApiException $e) {
+                switch ($e->getCode()) {
+                                        case 401:
+                        if ($retries < 1) {
+                            Authenticator::reauthenticate();
+                            continue 2;
+                        }
+                    default:
+                        $data = $this->apiClient->getSerializer()->deserialize($e->getResponseBody(), '\Secuconnect\Client\Model\ProductExceptionPayload', $e->getResponseHeaders());
+                        $e->setResponseObject($data);
+                        break;
+                }
+
+                throw $e;
+            }
+        }
+    }
 }